requirements.txt
requests
dnspython
networkx
pydot
